创建原始数据集对应的 milvus 数据库, 插入数据

+++ src/milvus/collection_drop.py
@@ -1,8 +1,10 @@
 from pymilvus import connections, Collection, utility
 
 connections.connect(alias="default", host="localhost", port="19530")
-collection_name = "rag_qwen"
+print("Collections: ", utility.list_collections())
+collection_name = input("Input Collection Name: ")
 collection = Collection(collection_name)
-utility.list_collections()
-utility.drop_collection(collection_name)
-print(f"Collection '{collection_name}' 已删除")
+yes = input("RU Sure?[yes/no]")
+if yes == 'yes':
+    utility.drop_collection(collection_name)
+    print(f"Collection '{collection_name}' 已删除")
